+++ find_duplicates/config.py
@@ -43,6 +43,7 @@ KEY_PUBLISHER_SOUNDEX = 'publisherSoundexLength'
 KEY_SERIES_SOUNDEX = 'seriesSoundexLength'
 KEY_TAGS_SOUNDEX = 'tagsSoundexLength'
 KEY_INCLUDE_LANGUAGES = 'includeLanguages'
+KEY_DISPLAY_LIBRARY_RESULTS = 'displayLibraryResults'
 KEY_AUTO_DELETE_BINARY_DUPS = 'autoDeleteBinaryDups'
 
 KEY_SHOW_VARIATION_BOOKS = 'showVariationBooks'